Renamed backend directoy Renamed backend directory and removed mock json data.
