--- ThriftUni/mock_backend/listingsData.json
@@ -1,22 +1,0 @@
-[
-  {
-    "id": "1",
-    "name": "Lorem Ipsum",
-    "details": "Details about product",
-    "price": "$10.05",
-    "status": "Pending",
-    "photos": [
-      "https://archive.org/download/placeholder-image/placeholder-image.jpg"
-    ]
-  },
-  {
-    "id": "2",
-    "name": "Lorem Ipsum",
-    "details": "Details about product",
-    "price": "$5.05",
-    "status": "Sold",
-    "photos": [
-      "https://archive.org/download/placeholder-image/placeholder-image.jpg"
-    ]
-  }
-]
